Checkpointing this code - set up to do the full 'mbedtls' handshake, but running out of stack space early in the process
There were other problems with this code -- like, I wanted to add a
field to the nopoll 'ctx' struct but couldn't reference the field I
added (a pointer to the mbedtls 'ssl' struct) for reasons mysterious.

Is possible the issue is including the '-lssl' lib at the same time as
the '-lmbedtls', resulting in bloatage somehow. But removing that
library reference would mean I'd need to replace a lot of stuff at
once..