Checkpointing - still some rough butchery in here, but the "websocket" sample compiles again. Still need to wire in the "mbedtls_ssl_write()" and "mbedtls_ssl_read()" functions
There is basically and ugly imbalance still between the "mbedtls"
initialization that are forced inside of the "nopoll_conn_tls_new()"
api, versus the "__nopoll_conn_new_common()" api it is supposed to
call. But we're still initializing the "libmbedtls" stuff as required,
and the websocket app that is linking with this modified version of
"libnopoll" is compiling again.

Next: wire in the actual "read" and "write" routines..