Hacked the "nopoll" (a.k.a. websocket) library to use our customized "mbedtls" library to establish a TLS 1.2 connection to the Currant backend server (must use "nopoll_conn_new()" api)
The "nopoll" library used to use the "ssl" library to create secure
connections, but this library didn't support TLS 1.2 - and the Currant
backend requires TLS 1.2.

The "mbedtls" library does support TLS 1.2, so the idea was to replace
"ssl" with "mbedtls"; but there were some difficulties and time pressure so
we decided to leave the "ssl" library references in place, but disable the
code that would allow a client to actual spin-up an "ssl" secured
websocket.

Instead (this is the hacky part), the "nopoll_conn_new()" api (which is
supposed to create an insecure websocket connection) was hijacked to call
into an mbedtls "shim" that spins up a secure TLS 1.2 connection with the
Currant backend server. The "nopoll" library then creates a websocket
connection on top of this secured socket in the usual way.

Also fixed a bug in the "nopoll" library's implementation of the websocket
protocol "pong" packet (it wasn't masking the packet etc).

Verified that we can talk to the Currant backend using Lester/Levy protocol
ping/pong packets.